Bare-bones code added
This commit introduces all of the classes I expect we'll need for our project.
(Undoubtedly, we will add, remove, and change some as we work on the project,
but this is what I foresee us needing as best I can look into the future.) I
haven't implemented any of the interesting methods in these classes, but I have
added extensive JavaDoc and many of the classes' member variables. By doing so,
I have effectively documented this project's architecture.

I'm doing this because I will be out of town for an entire month, but I want to
be able to hit the ground running writing this program when I get back. By
documenting the architecture now, I am making sure it will be as easy as
possible for me to become productive quickly next month.