Unsuccesful attempt to implement GitCaller
This commit implements most of what I think we need for the GitCaller.
Unfortunately, when the test I wrote in Main is run, it doesn't seem to matter
what string I put in as the path to the voldemort repository - jGit doesn't
seem to be opening the repository correctly. I'm going to investigate this more
later, but for now, the changes I've made are nontrivial enough to be worth
committing.

Also, this commit involved a significant change in the program's design.
Previously, my plan was to use the GitCaller to get a list of hashes for the
commits we're interested in, then process each hash individually to get the
other commit information (like author and date) that we want. However, it looks
vastly more efficient with jGit to get all of the information we can for each
commit all in one pass. So that's what we're doing now.