Sorting files into components for layers
Since most projects have dozens or hundreds of files, we want to
reduce the number of file-type entities we eventually display in
our visualization. There just isn't enough space to give each file
its own layer in the visualization without making the display
unreadable.

As a result, we sort each file into "Components," or logically
distinct entities in the target software's design. We do so using
a naive approach: The user gives a list of regexes representing
important directories in the software's directory structure, and
each file inside one of those directories is considered to be part
of that directory's component. This is flexible between different
projects and architectures and is relatively transparent and easy
to understand for the user (I think; we might need to research/test
this assumption).

This commit introduces the logic to get the directory regexes from
the user and to sort files into components. We will continue using
this information in future data processing steps.