Incomplete - Implementing visualization logic
This commit starts implementing the logic for producing the model
that will be used in the visualization. The algorithim in this
commit *is incorrect*, but I've done so much work on it so far
that's worth keeping that I thought it would be best to commit
my work so that it's backed up and so that I can return to it if
future experiments go even worse.