Deprecated 3D view and started 2D view
Since the 3D view is very CPU-hungry and relatively unintuitive to
use, we're replacing it with a couple of 2D views:

One of the new views will show the branching pattern of the history,
but not which files were modified in each commit.

CLicking and dragging over some of the commits in that view will open
a new view, which will show which files were modified in the selected
commits, but not the branching and merging structure.