Keybaord events.
I'm using a variation of the strategy pattern for handling keyboard
events in the TwoDimensionalView. The KeyboardEvent abstract class
specifies several event handler methods, and these methods are called
at the appropriate points from within TwoDimensionalView. The end
result is that any arbitrary keyboard buttons can be tied to any
arbitrary behavior in the UI, and instead of having to write
boilerplate code in each implementation of TwoDimensionalView,
the views can just create a list of KeyboardEvents and all the
bookkeeping is handled for them.