Playing sounds for commit authors
Reworked some of the backend to senf information about how many
commits each author made to the BranchView, letting sounds be assigned
based on who made the most commits.

The sounds don't quite work correctly right now. If a new commit is
highlighted that plays the same sound as the previous commit, it will
sometimes continue playing the sound instead of stopping and
restarting it, as it should.