Started algorithm for getting branches from git log --graph
The current algorithm for generating the graph visualization doesn't
work -- the graph shown by the visualization disagrees with the graph
produced by git log --graph. So to fix it, I'm going to just get the
commit and branch positions directly from git log --graph instead of
trying to create my own algorithm to show the graph.

This commit starts the algorithm to get the necessary data out of
git log --graph. It's untested and is missing an important function,
but it represents enough work that I don't want to risk losing the
progress I've made.