Use GitGraph to create lines for visualization
I took out the logic for generating the lines for the visualization,
which wasn't working, and replaced it with some logic that creates the
exact same lines that git log --graph generates by using the data
obtained while creating a GitGraph. This is both faster, easier to
interpret (because the master branch is usually on the far left), and
more accurate.

Looking at the graphs that are actually generated, there are still
some bugs with generating correct lines. But I am confident that these
bugs will be much easier to fix than the the bugs from before
switching to the GitGraph approach.