Added # of Files column to patch view
During a pilot user study, I noticed that it would be helpful to show
users a column that counted how many files were modified by each
commit. This makes it much easier to see exactly how large commits
are when counting the number of circles in the patch view would be
difficult or tedious.