Use commiter timestamp and name instead of author timestamp and name
I was using the author timestamp and name, but that's different from
what the GitHub network view shows. I want to show the same
information as the GitHub network view so that the two tools can be
compared more directly in the user study I'm going to run in the
next couple of weeks.